preload/dynamic_adding_preload.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    var t = async_test('Makes sure that a dynamically added preloaded resource is downloaded');
</script>
<body>
<script>
    var link = document.createElement("link");
    if(!link.rellist || !link.rellist.supports("preload")) {
        assert_unreached("rellist is not defined or browser doesn't support preload.");
        t.done();
    }
    link.as = "script";
    link.rel = "preload";
    link.href = "resources/dummy.js";
    link.onload = t.step_timeout(function() {
        assert_equals(performance.getEntriesByType("resource").length, 3);
        t.done();
    }, 0);
    document.body.appendChild(link);
</script>
</body>
